Bugfix : * correction du probleme d'affichage des elements par sous-reseau * les RzxComputer connaissent desormais le module network duquel ils sont issus. ce qui regle au passage le probleme de telechargement des icones.
git-svn-id: svn://localhost/qrezix/branches/qrezix--qt4@873 30024bfc-89fa-0310-95ff-ee513f6fc940

+++ qrezix/core/rzxconfig.cpp
@@ -509,9 +509,9 @@ void RzxConfig::loadRezals()
 }
 
 ///Retourne le nombre de sous réseaux
-int RzxConfig::rezalNumber()
+uint RzxConfig::rezalNumber()
 {
-	return global()->rezalSubnets.count();
+	return (uint)global()->rezalSubnets.count();
 }
 
 ///Retourne le rezal auquel appartient le QHostAddress
@@ -539,7 +539,7 @@ int RzxConfig::rezal(const QHostAddress& addr)
  */
 QString RzxConfig::rezalName(int rezalid, bool shortname)
 {
-	if(rezalid >= rezalNumber() || rezalid < 0)
+	if(rezalid >= (int)rezalNumber() || rezalid < 0)
 		return QString();
 
 	if(shortname)
